Implement workaround to mount standard DAGs
This enables mounting the DAGs on k8s as well as on docker compose.
If the standard DAGs were copied directly to the /opt/airflow/dags
directory, they'd get deleted once the PVC is mounted in the same
place. By copying them to the /tmp directory and then copy them
again to the opt/airflow/dags folder on startup, it is ensured that
they are always present there.